requirement.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
appnope==0.1.2
async-timeout==3.0.1
attrs==21.2.0
autopage==0.5.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.10.0
bitarray==1.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
cliff==3.10.0
cmd2==2.3.3
cryptography==36.0.1
cycler==0.10.0
cytoolz==0.11.0
debtcollector==2.4.0
debugpy==1.4.1
decorator==5.0.9
distfit==1.4.2
dogpile.cache==1.1.5
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
html5lib==1.1
idna==3.2
ipfshttpclient==0.8.0a2
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
iso8601==1.0.2
jedi==0.18.0
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
keystoneauth1==4.4.0
kiwisolver==1.3.2
lru-dict==1.1.7
lxml==4.6.3
matplotlib==3.4.3
matplotlib-inline==0.1.2
msgpack==1.0.3
multiaddr==0.0.9
multidict==5.1.0
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.21.2
openstacksdk==0.61.0
os-service-types==1.7.0
osc-lib==2.4.2
oslo.config==8.7.1
oslo.i18n==5.1.0
oslo.serialization==4.2.0
oslo.utils==4.12.1
packaging==21.3
pandas==1.3.3
parsimonious==0.8.1
parso==0.8.2
patsy==0.5.2
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prettytable==3.0.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py-solc-x==1.1.1
pycparser==2.21
pycryptodome==3.10.1
Pygments==2.10.0
pyparsing==2.4.7
pyperclip==1.8.2
pypickle==1.0.0
pyrsistent==0.18.0
python-cinderclient==8.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-keystoneclient==4.4.0
python-novaclient==17.6.0
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.2.1
redis==3.5.3
requests==2.26.0
requestsexceptions==1.4.0
rfc3986==2.0.0
rlp==2.0.1
scipy==1.7.3
semantic-version==2.9.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.2.1
statsmodels==0.13.1
stevedore==3.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
varint==1.0.2
wcwidth==0.2.5
web3==5.28.0
webencodings==0.5.1
websockets==9.1
wrapt==1.13.3
yarl==1.6.3
